refactor(store): 기존 더미데이터 삭제 및 store 통합, 세터함수 추가 #155
